move edittional funciton to AgileacceleratorCommon
